Switching to NuGet for dependency management.
NuGet is now being used for dependency management. It's a little bit nasty
because VS just isn't good at dealing with a project that targets multiple
platforms in the way WPF Converters does. This means that maintaining
references will be little bit cumbersome and will require manual editing
of the project file. But project references will rarely if ever change.

The intention is to switch to portable class libraries in a coming release,
so hopefully this problem just goes away.